wip: create TopicConfigure
